Move coverage_area_km2 to ecoscope.base.utils; silence scipy-stubs overload error
The buffer-corridor area helper is pure geometry with only core deps,
so it belongs with create_meshgrid in base.utils rather than inside the
summarize_df task module (resolves the placement TODO; no duplicate
exists elsewhere in the codebase).

etd_range: scipy-stubs' LowLevelCallable overloads no longer accept
numba's _CFunctionType, which broke the mypy CI job on every branch.
