+++ ecoscope/analysis/UD/etd_range.py
@@ -36,6 +36,7 @@ def __etd__(_, a):
     return np.exp(-((s / el) ** k)) * s ** (k - 2) / np.sqrt(s**2 - m**2)
 
 
+# scipy-stubs' LowLevelCallable overloads don't accept numba's _CFunctionType
 _etd = scipy.LowLevelCallable(__etd__.ctypes)  # type: ignore[call-overload]
