+++ ecoscope/base/utils.py
@@ -78,6 +78,33 @@ def create_meshgrid(
     return gs
 
 
+def coverage_area_km2(
+    gdf: gpd.GeoDataFrame,
+    swath_width_meters: float,
+    merged: bool,
+    area_crs: str = "EPSG:6933",
+) -> float:
+    """
+    Compute the ground area (km²) covered by buffering track segments.
+
+    Each segment is buffered by ``swath_width_meters / 2`` per side to form a
+    corridor of full width ``swath_width_meters``. When ``merged`` is True the
+    area of the union of all corridors is returned (overlaps counted once);
+    otherwise the per-segment corridor areas are summed.
+    """
+    if gdf.empty:
+        return 0.0
+
+    buffers = gdf.to_crs(area_crs).geometry.buffer(swath_width_meters / 2)
+
+    if merged:
+        area_m2 = buffers.union_all().area
+    else:
+        area_m2 = buffers.area.sum()
+
+    return area_m2 / 1e6
+
+
 def groupby_intervals(df: pd.DataFrame, col: str, intervals: pd.IntervalIndex) -> DataFrameGroupBy:
     """
     Parameters
